--- baselines/fjord/fjord/fjord/od/models/__init__.py
@@ -1,3 +1,0 @@
-from .resnets import ResNet18, ResNet34, ResNet50, ResNet101
-
-__all__ = ['ResNet18', 'ResNet34', 'ResNet50', 'ResNet101']
