+++ baselines/fjord/fjord/scripts/run.sh
@@ -8,9 +8,9 @@ for seed in 123 124 125; do
     echo "Running seed $seed"
 
     echo "Running without KD ..."
-    poetry run python simulation.py ++manual_seed=$seed |& tee $RUN_LOG_DIR/wout_kd_$seed.log
+    poetry run python main.py ++manual_seed=$seed |& tee $RUN_LOG_DIR/wout_kd_$seed.log
 
     echo "Running with KD ..."
-    poetry run python simulation.py +train_mode=fjord_kd ++manual_seed=$seed  |& tee $RUN_LOG_DIR/w_kd_$seed.log
+    poetry run python main.py +train_mode=fjord_kd ++manual_seed=$seed  |& tee $RUN_LOG_DIR/w_kd_$seed.log
 done
 popd
