+++ baselines/fjord/fjord/strategy.py
@@ -1,4 +1,4 @@
-from typing import Dict, List, Optional, Tuple, Union, OrderedDict
+from typing import Dict, List, Optional, Tuple, Union
 from functools import reduce
 from copy import deepcopy
 
@@ -15,9 +15,6 @@ from flwr.server.client_proxy import ClientProxy
 from flwr.server.strategy import FedAvg
 
 import numpy as np
-import torch
-
-from torch.nn import Module
 
 from fjord.utils.logger import Logger
 
@@ -38,26 +35,6 @@ def weighted_average(metrics: List[Tuple[int, Metrics]]) -> Metrics:
 
     # Aggregate and return custom metric (weighted average)
     return {"accuracy": sum(accuracies) / sum(examples)}
-
-
-def get_parameters(net: Module) -> List[np.ndarray]:
-    """
-    Get statedict parameters as a list of numpy arrays.
-    :param net: PyTorch model
-    :return: List of numpy arrays
-    """
-    return [val.cpu().numpy() for _, val in net.state_dict().items()]
-
-
-def set_parameters(net: Module, parameters: List[np.ndarray]) -> None:
-    """
-    Load parameters into PyTorch model.
-    :param net: PyTorch model
-    :param parameters: List of numpy arrays
-    """
-    params_dict = zip(net.state_dict().keys(), parameters)
-    state_dict = OrderedDict({k: torch.Tensor(v) for k, v in params_dict})
-    net.load_state_dict(state_dict, strict=True)
 
 
 def get_p_layer_updates(
